requirements.txt
web3
psycopg2-binary
click
pyyaml
uniswap-python
clickhouse_connect
